requirements.txt
streamlit
pandas
opencv-python
face_recognition
pillow
numpy
cryptography
fastapi
uvicorn
python-multipart
